+++ packages/experiments/src/Foundation.ts
@@ -1,7 +1,7 @@
-import { mergeStyleSets, getTheme, ITheme } from 'office-ui-fabric-react';
+import { mergeStyleSets, ITheme } from 'office-ui-fabric-react';
 import {
-  createComponent as create,
-  createComponentWithState as createWithState,
+  createStatelessComponent as foundationCreateStatelessComponent,
+  createComponent as foundationCreateComponent,
   IComponentOptions,
   IViewComponentProps,
   IStateComponent,
@@ -13,6 +13,7 @@ import {
 } from '@uifabric/foundation';
 export { IStateComponentProps } from '@uifabric/foundation';
 import { IProcessedStyleSet, IStyleSet } from './Styling';
+import { Customizations, CustomizableContextTypes, ICustomizations } from './Utilities';
 
 // Centralize Foundation interaction for use throughout this package. These convenience types provide types
 //  that are global for all of OUFR, such as ITheme and IProcessedStyleSet.
@@ -34,34 +35,48 @@ export type IStyleableComponentProps<TProps, TStyleSet> = IStyleableComponentPro
  */
 export type IThemedProps<TProps> = TProps & IThemedComponent<ITheme>;
 
+/**
+ * The shape of customizations within context.
+ */
+type IContextCustomization = { customizations: ICustomizations };
+
+// TODO: remove any if possible
 // tslint:disable-next-line:no-any
-const providers: IStylingProviders<any, any, ITheme> = {
-  getTheme,
-  mergeStyleSets
+const providers: IStylingProviders<any, any, any, IContextCustomization, ITheme> = {
+  mergeStyleSets,
+  getCustomizations,
+  CustomizableContextTypes
 };
 
 /**
- * A helper for Foundation's createComponent that automatically passes in constant types.
- * See Foundation's createComponent for more detail.
+ * A helper for Foundation's createStatelessComponent that automatically passes in constant types.
+ * See Foundation's createStatelessComponent for more detail.
  * @param {IComponentOptions} options
  */
-export function createComponent<
+export function createStatelessComponent<
   TComponentProps extends IStyleableComponent<TComponentProps, TStyleSet, ITheme>,
   TStyleSet extends IStyleSet<TStyleSet>,
   TStatics = {}
 >(
   options: IComponentOptions<TComponentProps, TStyleSet, IProcessedStyleSet<TStyleSet>, ITheme, TStatics>
 ): React.StatelessComponent<TComponentProps> & TStatics {
-  return create(options, providers);
+  return foundationCreateStatelessComponent<
+    TComponentProps,
+    TStyleSet,
+    IProcessedStyleSet<TStyleSet>,
+    IContextCustomization,
+    ITheme,
+    TStatics
+  >(options, providers);
 }
 
 /**
- * A helper for Foundation's createComponentWithState that automatically passes in constant types.
- * See Foundation's createComponentWithState for more detail.
+ * A helper for Foundation's createComponent that automatically passes in constant types.
+ * See Foundation's createComponent for more detail.
  * @param {IComponentOptions} options
  * @param {IStateComponent} state
  */
-export function createComponentWithState<
+export function createComponent<
   TComponentProps extends IStyleableComponent<TViewProps, TStyleSet, ITheme>,
   TViewProps,
   TStyleSet extends IStyleSet<TStyleSet>,
@@ -74,5 +89,22 @@ export function createComponentWithState<
     IProcessedStyleSet<TStyleSet>
   >
 ): React.StatelessComponent<TComponentProps> & TStatics {
-  return createWithState(options, providers, state);
+  return foundationCreateComponent<
+    TComponentProps,
+    TViewProps,
+    TStyleSet,
+    IProcessedStyleSet<TStyleSet>,
+    IContextCustomization,
+    ITheme,
+    TStatics
+  >(options, providers, state);
+}
+
+// TODO: remove any if possible
+// tslint:disable-next-line:no-any
+function getCustomizations(displayName: string, context: IContextCustomization): any {
+  // TODO: do we want field props? should fields be part of IComponentOptions and used here?
+  // TODO: should we centrally define DefaultFields? (not exported from styling)
+  const DefaultFields = ['theme', 'styles'];
+  return Customizations.getSettings(DefaultFields, displayName, context.customizations);
 }