+++ packages/experiments/src/components/Stack/StackItem/StackItem.test.tsx
@@ -36,4 +36,35 @@ describe('Stack Item', () => {
 
     expect(wrapper.find('.test').length).toBe(0);
   });
+
+  it('can handle having no children', () => {
+    const createEmptyStackItem = () => {
+      mount(
+        <Stack>
+          <Stack.Item />
+        </Stack>
+      );
+    };
+
+    expect(createEmptyStackItem).not.toThrow();
+  });
+
+  it('includes the classNames on both a StackItem and its child', () => {
+    const stackItemClassName = 'stackItemClass';
+    const childClassName = 'childClass';
+
+    const wrapper = mount(
+      <Stack>
+        <Stack.Item className={stackItemClassName}>
+          <span className={childClassName} />
+        </Stack.Item>
+      </Stack>
+    );
+
+    const child = wrapper.find('span').at(0);
+    const childClass = child.props().className;
+
+    expect(childClass).toContain(stackItemClassName);
+    expect(childClass).toContain(childClassName);
+  });
 });