+++ packages/office-ui-fabric-react/src/components/Layer/Layer.tsx
@@ -4,5 +4,6 @@ import { LayerBase } from './Layer.base';
 import { getStyles } from './Layer.styles';
 
 export const Layer = styled<ILayerProps, ILayerStyleProps, ILayerStyles>(LayerBase, getStyles, undefined, {
-  scope: 'Layer'
+  scope: 'Layer',
+  fields: ['hostId', 'theme', 'styles']
 });