+++ packages/styling/src/styles/theme.ts
@@ -1,19 +1,19 @@
-import { Customizations } from '@uifabric/utilities';
+import { Customizations, merge } from '@uifabric/utilities';
 import { IPalette, ISemanticColors, ITheme, IPartialTheme } from '../interfaces/index';
-import { ITypography } from '../interfaces/ITypography';
+import { ITypography, IPartialTypography, IFontVariant } from '../interfaces/ITypography';
 import { DefaultFontStyles } from './DefaultFontStyles';
 import { DefaultPalette } from './DefaultPalette';
 import { DefaultTypography } from './DefaultTypography';
 import { loadTheme as legacyLoadTheme } from '@microsoft/load-themed-styles';
 
-let _theme: ITheme = {
+let _theme: ITheme = createTheme({
   palette: DefaultPalette,
   semanticColors: _makeSemanticColorsFromPalette(DefaultPalette, false, false),
   fonts: DefaultFontStyles,
   isInverted: false,
   typography: DefaultTypography,
   disableGlobalClassNames: false
-};
+});
 let _onThemeChangeCallbacks: Array<(theme: ITheme) => void> = [];
 
 export const ThemeSettingName = 'theme';
@@ -108,30 +108,21 @@ export function createTheme(theme: IPartialTheme, depComments: boolean = false):
     ...theme.semanticColors
   };
 
-  let typography = DefaultTypography;
-  if (theme.typography) {
-    typography = {
-      families: { ...DefaultTypography.families, ...theme.typography.families },
-      weights: { ...DefaultTypography.weights, ...theme.typography.weights },
-      sizes: { ...DefaultTypography.sizes, ...theme.typography.sizes },
-      types: { ...DefaultTypography.types, ...theme.typography.types }
-    } as ITypography;
-  }
+  const typography = merge<ITypography>({}, DefaultTypography, theme.typography as ITypography);
+  const { variants } = typography;
 
-  const { types } = typography;
-  for (const typeName in types) {
-    if (types.hasOwnProperty(typeName)) {
-      const type = types[typeName];
-      if (type) {
-        if (type.fontFamily && typography.families[type.fontFamily]) {
-          type.fontFamily = typography.families[type.fontFamily];
-        }
-        if (type.fontSize && typography.sizes[type.fontSize]) {
-          type.fontSize = typography.sizes[type.fontSize];
-        }
-        if (type.fontWeight && typography.families[type.fontWeight]) {
-          type.fontWeight = typography.families[type.fontWeight];
-        }
+  for (const variantName in variants) {
+    if (variants.hasOwnProperty(variantName)) {
+      const variant = variants[variantName];
+
+      if (variant.fontFamily && typography.families[variant.fontFamily]) {
+        variant.fontFamily = typography.families[variant.fontFamily];
+      }
+      if (variant.fontSize && typography.sizes[variant.fontSize]) {
+        variant.fontSize = typography.sizes[variant.fontSize];
+      }
+      if (variant.fontWeight && typography.weights[variant.fontWeight]) {
+        variant.fontWeight = typography.weights[variant.fontWeight];
       }
     }
   }
@@ -192,11 +183,22 @@ function _makeSemanticColorsFromPalette(p: IPalette, isInverted: boolean, depCom
     buttonBackgroundChecked: p.neutralTertiaryAlt,
     buttonBackgroundHovered: p.neutralLight,
     buttonBackgroundCheckedHovered: p.neutralLight,
+    buttonBackgroundPressed: p.neutralLight,
     buttonBorder: 'transparent',
     buttonText: p.neutralPrimary,
     buttonTextHovered: p.black,
     buttonTextChecked: p.neutralDark,
     buttonTextCheckedHovered: p.black,
+    buttonTextPressed: p.neutralDark,
+    buttonTextDisabled: p.neutralQuaternary,
+    buttonBorderDisabled: 'transparent',
+    primaryButtonBackground: p.themePrimary,
+    primaryButtonBackgroundHovered: p.themeDarkAlt,
+    primaryButtonBackgroundPressed: p.themeDark,
+    primaryButtonBorder: 'transparent',
+    primaryButtonText: p.white,
+    primaryButtonTextHovered: p.white,
+    primaryButtonTextPressed: p.white,
 
     menuItemBackgroundHovered: p.neutralLighter,
     menuIcon: p.themePrimary,