+++ scripts/tasks/lint-imports.js
@@ -1,11 +1,37 @@
+const importKeywordRegex = /^import/gm;
+const importStatementGlobalRegex = /^import [{} a-zA-Z0-9_,*\r?\n ]*(?:from )?['"]{1}([.\/a-zA-Z0-9_@\-]+)['"]{1};.*$/gm;
+const importStatementRegex = /^import [{} a-zA-Z0-9_,*\r?\n ]*(?:from )?['"]{1}([.\/a-zA-Z0-9_@\-]+)['"]{1};.*$/;
+const pkgNameRegex = /^(@[a-z\-]+\/[a-z\-]+)\/|([a-z\-]+)\//;
 module.exports = function(options) {
   const path = require('path');
   const fs = require('fs');
   const chalk = require('chalk');
   const sourcePath = path.resolve(process.cwd(), 'src');
+  const nodeModulesPath = path.resolve(process.cwd(), 'node_modules');
+  const rushJsonPath = findRushJson(process.cwd());
+  const rootFolder = path.dirname(rushJsonPath);
+  if (!rushJsonPath) {
+    throw new Error('lint-import: unable to find rush.json');
+  }
 
+  const rush = JSON.parse(fs.readFileSync(rushJsonPath, 'utf8'));
+  const rushPackages = rush.projects.map(project => project.packageName);
+
+  const currentRushPackage = rush.projects.find(project => {
+    return path.normalize(project.projectFolder) === path.normalize(path.relative(rootFolder, process.cwd()));
+  }).packageName;
   // TestCode
   return lintSource();
+
+  function findRushJson(curDir) {
+    const potentialRushPath = path.resolve(curDir, 'rush.json');
+    if (fs.existsSync(potentialRushPath)) {
+      return potentialRushPath;
+    }
+    const parentFolder = path.dirname(curDir);
+    if (parentFolder === curDir) return undefined;
+    return findRushJson(parentFolder);
+  }
 
   function lintSource() {
     const files = _getFiles(sourcePath, /\.(ts|tsx)$/i);
@@ -69,12 +95,8 @@ module.exports = function(options) {
   function _evaluateFile(filePath, importErrors) {
     const fileContent = fs.readFileSync(filePath, 'utf8');
 
-    // Find imports.
-    const importKeywordRegex = /^import/gm;
-    const importStatementRegex = /^import [{} a-zA-Z0-9_,*\r?\n ]*(?:from )?['"]{1}([.\/a-zA-Z0-9_@\-]+)['"]{1};.*$/gm;
-
-    const importKeywords = fileContent.match(importKeywordRegex);
-    const importStatements = fileContent.match(importStatementRegex);
+    const importKeywords = fileContent.match(importStatementGlobalRegex);
+    const importStatements = fileContent.match(importStatementGlobalRegex);
 
     importErrors.totalImportKeywords += importKeywords ? importKeywords.length : 0;
     importErrors.totalImportStatements += importStatements ? importStatements.length : 0;
@@ -87,7 +109,7 @@ module.exports = function(options) {
 
     if (importStatements) {
       importStatements.forEach(statement => {
-        const parts = new RegExp(importStatementRegex).exec(statement);
+        const parts = importStatementRegex.exec(statement);
 
         if (parts) {
           _evaluateImport(filePath, parts[1], importErrors);
@@ -97,21 +119,36 @@ module.exports = function(options) {
   }
 
   function _evaluateImport(filePath, importPath, importErrors) {
+    let fullImportPath;
     if (importPath.indexOf('.') === 0) {
       // import is a file path. is this a file?
-      const fullImportPath = _evaluateImportPath(path.dirname(filePath), importPath);
-
-      // Does this file path exist?
-      if (!fullImportPath) {
-        console.log(`DOESNT EXIST!!! ${importPath}`);
-      } else {
-        if (fs.statSync(fullImportPath).isDirectory()) {
-          const pathInvalid = importErrors.pathInvalid;
-          const relativePath = path.relative(sourcePath, filePath);
-          pathInvalid.count++;
-          pathInvalid.matches[relativePath] = importPath;
-        }
+      fullImportPath = _evaluateImportPath(path.dirname(filePath), importPath);
+    } else {
+      const pkgNameMatch = importPath.match(pkgNameRegex);
+      if (pkgNameMatch === null) {
+        // This means the import does not adhere to what we are looking for (usually import * from 'react';, which
+        // we would skipping linting for.
+        return;
       }
+
+      const pkgName = pkgNameMatch[1] || pkgNameMatch[2];
+
+      // we don't evaluate imports of non rush packages
+      if (rushPackages.indexOf(pkgName) === -1) return;
+
+      if (pkgName === currentRushPackage) {
+        const importPathWithoutPkgName = importPath.substring(pkgName.length + 1 /* 1 is for '/' */);
+        fullImportPath = _evaluateImportPath(process.cwd(), './' + importPathWithoutPkgName);
+      } else {
+        fullImportPath = _evaluateImportPath(nodeModulesPath, './' + importPath);
+      }
+    }
+
+    if (!fullImportPath || fs.statSync(fullImportPath).isDirectory()) {
+      const pathInvalid = importErrors.pathInvalid;
+      const relativePath = path.relative(sourcePath, filePath);
+      pathInvalid.count++;
+      pathInvalid.matches[relativePath] = importPath;
     }
   }